+++ src/main/java/ceos/backend/global/util/ParsedDurationConvertor.java
@@ -3,14 +3,13 @@ package ceos.backend.global.util;
 import ceos.backend.global.common.dto.ParsedDuration;
 import lombok.RequiredArgsConstructor;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.List;
 
 @RequiredArgsConstructor
-public class ParsingDuration {
+public class ParsedDurationConvertor {
     private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd HH:mm:ss");
     private static final DateTimeFormatter yearDateSlashFormmatter = DateTimeFormatter.ofPattern("yyyy/MM/dd");
     private static final DateTimeFormatter yearDateDotFormmatter = DateTimeFormatter.ofPattern("yyyy.MM.dd");
